requirements.txt
pip==23.1.2
django
psycopg2-binary
djangorestframework
python-decouple
django-filter
django-cors-headers
dj-database-url
whitenoise
drf_yasg
celery
redis
djangorestframework-simplejwt
django-crontab
dj_database_url
loguru
django-celery-beat
gunicorn
pytz
twilio

